requirements.txt
transformers
torch
Pillow
timm
requests